Move ThreadNotify into local_pool.rs
